Remove old test files
