+++ test/util.ts
@@ -1,12 +1,12 @@
 import { Firestore } from '@google-cloud/firestore'
 import crypto from 'crypto'
-import admin from 'firebase-admin'
+import admin, { ServiceAccount } from 'firebase-admin'
 // because this file path will be dist_test/test/util.js after tsc
-import serviceAccount from '../../firebase_secret.json'
+import serviceAccount from '../firebase_secret.json'
 
 export const initFirestore = () => {
   admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount),
+    credential: admin.credential.cert(serviceAccount as ServiceAccount),
   })
   const firestore = admin.firestore()
   firestore.settings({ timestampsInSnapshots: true })
